Add freezable Clock as a test helper
In order to be able to write tests that depend on time, it is useful to
be able to control time! Rather than invent a time machine and send the
entire computer to the right point in time, we simply create a mock
Clock which, in the test env only, allows us to "freeze" time, either at
the current time, or at the specified time.

In order to support this we also add a Schema, which extends the Ecto
schema, and allows our models (Currently only Snippet) to respect our
time traveling abilities.

Note that the Clock only compiles freezable methods in the configured
env (currently test) to prevent problems from occuring in production.
Because of this we actually have an un-testable code path, but since
it's simply a defdelegate to DateTime.utc_now() it shouldn't be an
issue.