Resolve stats error/refactor stats
This started out as "why are we getting an internal server error
sometimes".
It was because as part of merging the database results into the monthly
buckets one of the results would occasionally be for a month that was
outside of our bucket range. This was because we were truncating on the
postgres side, and then doing TZ conversions on the app side.

This PR resolves the issue by changing the queries so that we're
calculating the month buckets on the postgres side in the time zone the
application is running in. This means that DB results can simply be
converted from the DB structure, which looks like this:
[{july-2020, private: 2}, {july-2020, public: 3}]
to the structure we want, which looks like this:
[{july-2020, {3,2}}]
without any further time zone mucking about by the application.

We also solve another "key not in bucket map" issue by generating
12-months worth of buckets, ensuring that any results that appear 364
days before now() have a bucket to be placed into, and then deleting
the oldest bucket and any results placed into it (Ensuring our desired
"this month and previous 11 months" data, all of which is for whatever
"month" means in the application-specified timezone.)