Fix failing tests, clean up test a bit
Turns out that when you're testing against datetimes in the application
TZ, and freezing time to do the test, you really shouldn't call NOW() in
your SQL query.

There is a potential *minor* performance degredation here as we're
taking the NOW() out of the DB engine and instead passing the utc_now()
from the application into the DB query, forcing the DB to parse it.
Ultimately this is so minor I chose to disregard it. But I did feel a
little bad. :-)